+++ spec/mailtrap/mail_spec.rb
@@ -52,29 +52,6 @@ RSpec.describe Mailtrap::Mail do
 
         it { is_expected.to eq(expected_headers) }
       end
-
-      context 'when reply-to is added in varying formats' do
-        before do
-          message.reply_to = 'Reply To <reply-to@railsware.com>'
-        end
-
-        it 'excludes reply-to from custom headers' do
-          expect(headers).not_to have_key('Reply-To')
-        end
-      end
-
-      context 'when custom header and reply-to variants are present' do
-        before do
-          message.header['Reply-To'] = 'Reply To <reply-to@railsware.com>'
-          message.header['REPLY-TO'] = 'Upper Case <upper-reply-to@railsware.com>'
-          message.header['reply-to'] = 'Lower Case <lower-reply-to@railsware.com>'
-          message.header['X-Special-Domain-Specific-Header'] = 'SecretValue'
-        end
-
-        it 'keeps only custom headers and strips all reply-to header variants' do
-          expect(headers).to eq('X-Special-Domain-Specific-Header' => 'SecretValue')
-        end
-      end
     end
 
     describe '#reply_to' do
@@ -84,18 +61,6 @@ RSpec.describe Mailtrap::Mail do
 
       it 'maps reply-to to the structured field' do
         expect(mail.reply_to).to eq(email: 'reply-to@railsware.com', name: 'Reply To')
-      end
-
-      context 'when reply-to header variants are present' do
-        before do
-          message.header['Reply-To'] = 'Reply To <reply-to@railsware.com>'
-          message.header['REPLY-TO'] = 'Upper Case <upper-reply-to@railsware.com>'
-          message.header['reply-to'] = 'Lower Case <lower-reply-to@railsware.com>'
-        end
-
-        it 'maps the reply-to value to the structured field' do
-          expect(mail.reply_to).to eq({ email: 'lower-reply-to@railsware.com', name: 'Lower Case' })
-        end
       end
     end
 
@@ -199,31 +164,16 @@ RSpec.describe Mailtrap::Mail do
       its(:template_variables) { is_expected.to eq('first_name' => 'John') }
     end
 
-    %i[from to cc bcc].each do |header|
+    %w[from to cc bcc reply-to].each do |header|
       context "when '#{header}' is invalid" do
         let(:message_params) { super().merge(header => 'invalid email@example.com') }
 
         it 'raises an error' do
           expect { mail }.to raise_error(
             Mailtrap::Error,
-            "failed to parse '#{header.capitalize}': 'invalid email@example.com'"
+            "failed to parse '#{Mail::Field::FIELD_NAME_MAP[header]}': 'invalid email@example.com'"
           )
         end
-      end
-    end
-
-    context "when 'reply-to' is invalid" do
-      let(:invalid_reply_to) { 'invalid email@example.com' }
-
-      before do
-        message.header['Reply-To'] = invalid_reply_to
-      end
-
-      it 'raises an error' do
-        expect { mail }.to raise_error(
-          Mailtrap::Error,
-          "failed to parse 'Reply-To': 'invalid email@example.com'"
-        )
       end
     end
   end